[sys] OutgoingDirectory checks for right threading usage
This CL check that the dispatcher servicing the outgoing directory
is the same one as the one that created it.

RELNOTES: If you get crashes messages like
"|component::OutgoingDirectory| is thread-unsafe", that's a sign that
the outgoing directory is being used or destroyed in parallel to the
async dispatcher thread(s). To fix the crash, always manage/use the
outgoing directory object on the dispatcher thread, and the dispatcher
must also be single-threaded. If you are using a |fdf::Dispatcher|, then
the outgoing directory object must be managed/used from a driver
dispatcher task, and the driver dispatcher must be synchronized.

Bug: 113997
Run-All-Tests: True

Change-Id: I4bf3ec61315f95a053d4c77622d84d60bcec5aee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/750532
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>