[network/test] Relax Routes Watcher test asserts
Some of the `fuchsia.net.routes/WatcherV{4,6}` tests are prone to
flakiness because they unnecessarily require the Netstack to be fully
initialized before the watcher client is created.

This CL updates such tests so that they pass even if the Watcher is
installed before the Netstack finishes initialization.

Multiply: netstack-routes-fidl-test

Bug: 121918
Change-Id: I02ab6e0575ff4654957ff91fdcad145ad7031c9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803889
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>