[bazel_sdk] Remove fuchsia_only_target()
The fuchsia_only_target() function is not documented and seems to be
used internally to return a value for `target_compatible_with` argument
in rule invocations.

However, using a value of `[ "@platforms//os:fuchsia" ]` is both simpler
and equivalent to express that a target can only be built in Fuchsia
compatible build configurations.

Moreover, the fuchsia_only_target() implementation could never trigger
the error message of fuchsia_select() since it called it with a
"//conditions:default" condition label, which would always be matched
anyway, so there was no benefit in using it.

The function is no longer used internally by @rules_fuchsia, and is
not being used by downstream clients, so can be removed safely.

Bug: None

Change-Id: Ib7b06138fbaa7c373775f52deb1b8a39bf53b374
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804948
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>