[timekeeper] Adds a loop pause for suspendable sources
Adds a pause in the sampling loop when the time source can be suspended,
as in the case of pull sources. The old loop was adequate for push
sources since it would block until the push source sends a report.

In case of pull sources, having no pause here means that we will spin
and pull time information constantly from time sources. This is
exactly not what we want with pull source. Instead, we want to limit
the number of times that timekeeper calls into the time source to pull
time from, and minimize the length of time that the time source needs
to run.

The pause was added at the top level (instead of e.g. at the level of
the individual source) to ensure that all work related to time sampling
is completed before the clock maintenance loop goes to sleep.

Bug: 121920
Change-Id: Ic539874d365973145740389ea54437c0920383f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804656
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>