[diagnostics] Add shards to support soft transitions
Add a new core shard for wlancfg pointing to the future source of
the diagnostics persistence capability. This facilitates a soft
transition of certain out of tree builds. In a soon-to-arrive
change "new" will become the current shard.

Also add empty shards, that will support transitioning a route
for networking. These shards will then be filled in with an
appropriate route to make an atomic change for all builds.

Bug: 120916
Change-Id: I9932aeab4ccbaebf566f4846c1fbd3dddbac4dc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803367
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>