[usb] usb-peripheral not accounting uninitialized content
In some situations we're strncpy()ing from data stored in a std::string
using std::string::c_str() in the amount of std::string::size(). This
only copies the string contents (trailing null excluded). If there
happens to be junk in the destiation, it will show up in the output.

Bug: 121142
Change-Id: I80e8586c0a35f1e7271892109cc0b91ba899541b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803250
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>