[phys][gcc] Add PANICs to enum class switch-LUTs.
In a few different places, there were functions/lambdas which
basically acted as LUTs for enum class values implemented using switch
statements.  Even though all of the define values for the enum-classes
were covered by case statements, the compiler can still complain that
a function may not return a value depending on the compiler options
which were chosen.

Given the potential for uninitialized variable and/or memory
corruption, a demand from the compiler to define behavior of
"impossible" class enum value does not seem unreasonable.

So, in the places that the compiler was complaining, add calls to
ZX_PANIC in the case an unrecognized value is found. PANIC is flagged
as no_return, so this has the effect of suppressing the compiler
warning that "all paths must return a value".

Fixed: 122177
Change-Id: If21aab40e7a79f792fe2dcdcf165ac6ad518473a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806228
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>