[scenic][test] Fail louder
The test fixture ignored errors in a FIDL call to scene manager,
which results in a confusing failure mode in tests.

This makes the failure a bit louder, in the hope it will be
recognized for what it is in the future.

Bug: 122116
Change-Id: I38e904700ee2d047018f2551a2a2e40acfad5c93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807449
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>