[media/ffmpeg] fix broken profile dependencies
During the transition to product-specific profiles, we want the
'default' and 'max' profiles to be the same. 'max' is always used,
and we want to use 'default' instead when a product doesn't
specify a profile.

The BUILD.gn that defines the default and max profiles was using
deps in a way that was not effective (max ended up empty). This CL
makes both default and max depend on a common group target that
has the right deps.

Test: build change pertaining to ffmpeg roll only

Change-Id: I955bd41df8e72b0f9e43d179c4dcb372e06df428
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807607
Reviewed-by: Tom Bergan <tombergan@google.com>
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>