[ui-input] Enable auto gain, clean up test comments
Enables auto gain behavior in the light sensor. Test comments
were out of date and showed values calculated with calibration, when
calibration is currently only used for the lux and cct values.

This also updates the default gain value for the light sensor so that it
matches the initial configuration for auto gain.

Fixed: 116095, 100661
Test: fx test -o input_pipeline_lib_tests, manual test on sherlock,
astro, nelson

Change-Id: I97769aaccbd86f4eb602f0f5d35f83c82daed467
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768966
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>