[netstack2] handles route interrupts in order
Update the `routeWatcherEventLoop` to select on interrupts in a
deterministic order.

The original implementation would evaluate cases in a psedorandom
fashion (as is the behavior for Golang `select` statements). However,
this leads to pseudorandom behavior observed by clients of the watcher.
Clients may observe routes as either `Added`/`Existing` depending on the
order of case evaluation.

This CL uses a single channel to hold all interrupt types, which ensures
that the interrupts are handled in order.

Multiply: netstack-routes-fidl-test

Fixed: 121918
Change-Id: Ia6e33902a73130ea135ea71c021be5921627a603
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804929
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>