[rtl8111] Rewrite in C++
The motivation for this was to migrate FIDL from Banjo to PCI, which
requires the use of a C++ FIDL client. When I changed the file extension
from .c to .cc I found that extensive changes were needed to even get it
to compile (mainly because of gotos), so I went ahead and rewrote it in
idiomatic C++.

Disclaimer: We are not sure what hardware this driver runs on, and
consequently this CL has not been tested.

Bug: 99391
Change-Id: I9ca1c5dcf869cd433eb2432d8b6e8168b4bc600e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727017
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>