[platform-device] Always offer FIDL protocol
Always offer the FIDL protocol. This was not possible to do back when
setting an outgoing directory required the device to isolate.

This should make it easier to migrate all clients to using platform
device FIDL.

Update the `if` statements to a `switch` statement to ensure we check
all variants of the enum.

Bug: 110796
Change-Id: I130b03b84f34164e10fe632a50b6d17000a6625b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807228
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>