[services] Ignore NOT_DIRECTORY when opening service source directories
FilteredServiceProvider used the flags passed to Open by the client
as-is to open the source service directory. This failed when the flags
had NOT_DIRECTORY, e.g. when connecting directly to a protocol
within the service.

The provider now ignores the NOT_DIRECTORY flag if given by the client,
and always opens the service directory with the DIRECTORY flag.

Bug: 120673
Change-Id: I212b0e3f250c3d7616b0caa4a4977fd8cae4d5e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796602
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>