[netstack3] Introduce weakly referenced device IDs
Weak device IDs are IDs that may be held by things that do not make any
claim about the live-ness of a device. For example, sockets may hold a
weak reference to the device ID at rest since sockets do not claim that
the device will be alive for the lifetime of the socket. Instead,
sockets only need to ensure that a device is alive when attempting to
send packets.

Bug: 100759
Change-Id: Ic0816a35109383fe1e255bb22930817b7d15f601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807730
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>