[trace] Enable Tracing via common_minimum_eng AIB
When building eng builds, trace manager is added as a child to core so
that trace providers can connect to it. This allows us to remove
tracing in non eng builds by excluding it in the core shard as opposed
to the current behaviour which is to always including it in the core
shard as a child and simply not provide the trace_manager package,
which works, but causes verbose logs spew when the package is requested
but not found.

This is a stepping stone -- this cl fixes enough optional routes that
building without the core shard is only works for non-eng core right
now. Non-eng Workstation needs additional work. We'll need to do a
through cleaning of the fuchsia.tracing.provider.Registry capability
routing to removing usages of it that are non optional.

Bug: 112433
Change-Id: Ife8a81e727f26a6d978ccae9ecc13b08d97ba78d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779082
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Size-Review: Gopi Krishna Chitluri <gopichitluri@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>