[pistress] Update pistress to include NO_INHERIT profiles.
Update the PI stress test to also include fair profile threads which
do not have inheritable profile.

Also, stop passing periods which are not equal to deadlines.  Period
was effectively removed from the deadline scheduler parameters, and
should eventually be removed from the profile interface, so there is
not much point in us passing it.

Finally, add pistress to the zircon debugging bundle so that it will
get built by the bots, and hopefully will not end up breaking due to
changes in other parts of the system.

Change-Id: I0d0489669d2999899b88a16ecfd34078141aeaa1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793566
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>