[kernel][vm] Debug compressor
Adds a debug compressor attached to the page queues that, when enabled,
will immediately compress a small percentage of pages added to the page
queues. This is intended for use in test runs where, due to fast
execution and lack of memory pressure, compression would not otherwise
be exercised.

Bug: 60238
Change-Id: I15fac84d5401b1f0037f2100b08151c29e9245f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/808367
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>