[kernel] Remove bad dependency in validate-kernel-symbols rule
The action uses the link_output_rspfile() target in deps to
deduce the input file and gets it into the depfile, so this
`inputs` list is superfluous and erroneous in some variant
configurations. However, that's only true when the script
is fixed to actually emit the right things in its depfile.

Change-Id: Id344fc0144c4b4107b67117d358849e1143381c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809528
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>