[i2c][amlogic] Make aml-i2c read I2C channel metadata to find bus base
Make aml-i2c read the I2C channel metadata to see if only devices on a
single bus are present. If so, this bus ID is used as the bus base
reported by the driver, which allows multiple aml-i2c instances to
coexist. This way board drivers can be moved to the single-bus model
individually.

Bug: 120971
Test: I2C devices work on Nelson
Test: fx test aml-i2c-test
Change-Id: I37e4d6f0783b9b881593ae201b01b4cc4220e4b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804990
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>