[netstack] Add network-with-tracing-basic
Currently the network realm that has tracing enabled is configured to
include advanced features, which is the default. To make it easier to
trace on products that require the basic network realm, add a separate
variant of the network package for that and update the documentation to
explain it.

Change-Id: Id0b175604560db9675d12cfa4ffc2108ec75dcdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810047
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>