[phys] Update ElfImage API for clarity (and for a few bugs)
Previously, out-of-process loading was expected to use ElfImage::Load()
followed by ElfImage::SetLoadAddress() giving the actually desired
virtual address. This led to a bug in which Load() assumes that the
final load address is to be where the image ends up in the current
address space: it tells gSymbolize about this, which prints the wrong
load address. Accordingly, we update ElfImage::Load() to optionally take
a load address up front.

Out-of-process loading also needs the address of the memory image in the
current address space, however, which is abstracted away by the current
interface. Previously, the only way to access this would be to call
`LoadAddress()` after `Load()`, but before `SetLoadAddress()` to specify
the address under the new mapping: this is awkward - and also is no
longer feasible with the change described above. Accordingly, we now
provide an accessor for the image in the current address space as well.

Further `elfldltl::LoadInfo load()` has been renamed to `load_info()` to
disambiguate it from `Load()`.

Bug: 91187
Change-Id: I87ae9bc92e9f7ad642be4846b3ac0122bba9c7e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798902
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>