[ktrace_provider] Don't spin if we get a trace record with a zero header
Before this change, if ktrace_provider encountered a trace record with
a zero header, it would get into an infinite loop.  The trace record
would be treated as having zero size.  StopKTrace() and
ReadNextRecord() would try to skip over the record, but that's not
possible because we don't know how large the record is.
ReadNextRecord() would advance 'current_' by zero each time and so
would return the same pointer repeatedly.

This manifested itself when the following change was applied, because
that would cause the kernel to sometimes generate trace records with a
zero header when tracing was stopped concurrently with a trace record
being emitted:
https://fuchsia-review.googlesource.com/c/fuchsia/+/804268
"[kernel][ktrace] Fix race condition in ktrace record creation"

Bug: 122181
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/microbenchmarks_test
  with CL above applied
Change-Id: Iea42a02bd8ca898cd0b755960dd19389b1cddcd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809688
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>