[events] Disambiguate between multiple "use" statements with the same name"
In order to fix a bug which causes a component to not be able to listen to
more than one CapabilityRequested events.
This CL modifies routing so that it takes a Use decl or
a name. In the case of a use decl, it compares against the entire
decl when performing routing, rather than just the name.
Note that tools such as scrutiny and the routing lib tests
still use names rather than use decls.

Run-All-Tests: true

Bug: 121625
Change-Id: Id2fbc70fb6c9634896037d6284a06add257c59f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800163
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Gary Bressler <geb@google.com>