[media][codec] Teardown loop before CodecFactoryApp
If CreateChild or OpenExposedDir callbacks in ForwardToIsolate are outstanding when CodecFactoryApp is destroyed, then the failure_func callback will remain alive after the CodecFactoryApp. When the loop is shutdown, that will cause the FIDL bindings to get errors and failure_func to be called. For CodecFactoryApp::DiscoverMagmaCodecDriversAndListenForMoreAsync, this will cause TeardownMagmaCodec to be called, which will cause a UAF of the CodecFactoryApp.

To prevent this, always shutdown the loop before destrying the
CodecFactoryApp.

Bug: 121693
Change-Id: Ib1e4980c282b2ec0159400971baefde05e14f411
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801427
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>