[kernel][vm] Propagate memory priorities from VMARs
Adds a mechanism to VMARs to allow them to have a MemoryPriority
applied to them and propgated to subvmars, VMOs and the aspace.
Application of the MemoryPriority is currently hooked up to the
ALWAYS_NEED hint and replaces the latency_sensitive setting.

Unlike the previous latency_sensitive tag, which would be propagated as
a fuse that could never be reset, the memory priority tag propagates
both setting and unsetting.

Bug: 101641
Change-Id: I24f3b35e65743743ee9e8de81c4f184328927573
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806648
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>