[fbl][test] Fix an issue with running tests out of order.
Fix an issue with the Basic RefPtr test which could cause it to fail
if the tests were executed out of order.

To summarize: There was a helper class (RefCallCounter) which was used
to count the number of times a given RefPtr had AddRef and Release
called, and also kept track of the global number of times that RefPtr
instances were destroyed.

The `destroy_count` was tracked with a static member of the class, and
was properly statically initialized to 0.  The trouble is that only
the Basic test cared what its eventual value was, and it assumed that
the value always started at zero.  If other tests (who also made use
of the `RefCallCounter` class) had run first, it meant that the Basic
would be run after a number of `RefCallCounter` instances had already
been created and destroyed, invalidating the "starts at zero"
assumption.

Fix this issue by putting running all the tests in a test fixture, and
have that fixture reset the static bookkeeping during the fixture's
Setup hook.

Fixed: 92566
Change-Id: I5f816f0058be3aac0baea0248e635209e487ccda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811618
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>