[netstack3] Choose netstack via assembly config
Add a configuration arg to choose which of Netstack2 and Netstack3 is
used in the product image at product assembly time based on platform
config or board feature.

Add a new "core-with-netstack3" product to be used in CI builds and
temporarily propagate use of the "use_netstack3" GN arg to allow
overriding the choice of netstack for the existing "core" product. The
latter portion will be reverted once the Netstack3 CI builders are
migrated to use the new product.

Bug: 121811
Change-Id: Ic02dbfe90db9bbe0f1ed1d9c658bd3821fd2f631
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810887
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>