[netstack3] Store `user_timeout_until` for TCP retrans timers
... instead of `start`. As pointed out in this [comment], this is more
consistent with other protocols in netstack3, so that changing a socket
option doesn't take effect immediately but rather at the next time a
timer needs to be created. This is not exactly what Linux does, but it
is also unlikely that an application can depend on this and tell
difference.

[comment]: https://fuchsia-review.git.corp.google.com/c/fuchsia/+/808350/comment/4d3703e4_4ebb2ce2/

Change-Id: I2812a1e23884f6c3e5686fb6ffff7a5cf7ba577d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810598
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>