[pixel][test] Robustifies pixel tests
The tests as written were sensitive to the exact
nuance of colors used to examine the scene
histogram.

The new code matches the colors approximately, and does not
rely on exact color matching. Not quite a SNR calculation,
but enough for this test.

I also added a histogram printout as part of the test logs,
which hopefully offers more insight into why the tests might
fail if (when) they do.

Bug: 112452
Change-Id: Ibd630a776bb93b78c5e9f9be03a13169a7e0a723
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811967
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Fuchsia-Auto-Submit: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>