[fdf] Wait for DFv2 node to be removed completely
This CL addresses two issues related to shutting down DFv2 nodes. The
first issue is that DFv2 nodes do not confirm that its associated
component has been fully removed by the component manager before
allowing another component to be created with the same configuration. If
a new node is added after the old one is removed and the component
manager hasn't had time to remove the old component then it may complain
about duplicate components when the new node attempts to create a new
component.

In this CL, nodes will now send a `DestroyChild` request to the
component manager during the node removal process and will only be
considered fully removed when the FIDL request gets a response. This
ensures that the associated component has been fully removed.

The second issue is that DFv2 nodes can be added after another node with
the same name has been requested to be removed but before it has been
fully removed. The driver runner will complain that duplicate nodes
exist when attempting to add the new node even if the old node is in the
middle of being removed.

In this CL, DFv2 nodes will now check if the duplicate node is in the
middle of being removed and create a callback to create the new child
node once the old node is removed.

Both these issues are race conditions and so the test added in this CL
may not fail every time. It usually failed every 1 out of 3 tests.

Multiply: rebind_test
Bug: 121991
Bug: 121990
Change-Id: I4ad31e5ef32d1822c57c96be1c9b636eb192b4c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800764
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>